Refactor of the RobotImporter Gem files from the Utils directory. (#1070)
* Remove unused functions.
* Remove unnecessary dependencies from Utils.
* Split RobotImporterUtils file.
* Split RobotImporterUtils into smaller, thematically coherent files.
* Split Utils/SourceAssetsStorage into smaller, thematically coherent files.
* Change a generic Sdf folder name.
* Format changed files with clang-format.

(cherry picked from commit d7a751a8b5303d431eb62379d5519f8241a12db7)
