+++ Gems/ROS2RobotImporter/Code/Source/RobotImporter/RobotImporterWidget.h
@@ -24,7 +24,6 @@
 #include <AzCore/std/parallel/thread.h>
 #include <AzCore/std/smart_ptr/shared_ptr.h>
 #include <RobotImporter/FixURDF/URDFModifications.h>
-#include <RobotImporter/Utils/RobotImporterUtils.h>
 #include <RobotImporter/xacro/XacroUtils.h>
 
 #include <AzToolsFramework/API/ToolsApplicationAPI.h>
@@ -82,7 +81,7 @@ namespace ROS2RobotImporter
         bool m_copyReferencedAssets{ false };
 
         /// mapping from unresolved URI to asset source
-        Utils::ReferencedAssetMap m_referencedAssetMap;
+        Assets::ReferencedAssetMap m_referencedAssetMap;
 
         AZStd::shared_ptr<AZStd::thread> m_copyReferencedAssetsThread;
         AZStd::unique_ptr<SdfPrefabMaker> m_prefabMaker;
