+++ Gems/ROS2RobotImporter/Code/Source/RobotImporter/URDF/ArticulationsMaker.cpp
@@ -9,8 +9,8 @@
 #include "ArticulationsMaker.h"
 #include <AzCore/Component/EntityId.h>
 #include <AzToolsFramework/Entity/EditorEntityHelpers.h>
+#include <RobotImporter/Queries/SdfQueries.h>
 #include <RobotImporter/Utils/DefaultSolverConfiguration.h>
-#include <RobotImporter/Utils/RobotImporterUtils.h>
 #include <RobotImporter/Utils/TypeConversions.h>
 #include <Source/EditorArticulationLinkComponent.h>
 
@@ -52,7 +52,7 @@ namespace ROS2RobotImporter
             const sdf::JointAxis* jointAxis = joint->Axis();
             if (jointAxis != nullptr)
             {
-                jointCoordinateAxis = Utils::TypeConversions::ConvertVector3(jointAxis->Xyz());
+                jointCoordinateAxis = Utils::SDFormat::TypeConversions::ConvertVector3(jointAxis->Xyz());
                 quaternion = jointCoordinateAxis.IsZero()
                     ? AZ::Quaternion::CreateIdentity()
                     : AZ::Quaternion::CreateShortestArc(AZ::Vector3::CreateAxisX(), jointCoordinateAxis);
@@ -103,9 +103,9 @@ namespace ROS2RobotImporter
             AZStd::max(articulationLinkConfiguration.m_solverVelocityIterations, Utils::DefaultNumberVelSolver);
 
         articulationLinkConfiguration.m_mass = inertial.MassMatrix().Mass();
-        articulationLinkConfiguration.m_centerOfMassOffset = Utils::TypeConversions::ConvertVector3(inertial.Pose().Pos());
+        articulationLinkConfiguration.m_centerOfMassOffset = Utils::SDFormat::TypeConversions::ConvertVector3(inertial.Pose().Pos());
 
-        if (!Utils::TypeConversions::ConvertQuaternion(inertial.Pose().Rot()).IsIdentity())
+        if (!Utils::SDFormat::TypeConversions::ConvertQuaternion(inertial.Pose().Rot()).IsIdentity())
         { // There is a rotation component in URDF that we are not able to apply
             AZ_Warning(
                 "AddArticulationLink", false, "Ignoring URDF/SDF inertial origin rotation (no such field in rigid body configuration)");
@@ -130,9 +130,9 @@ namespace ROS2RobotImporter
 
         constexpr bool getNestedModelJoints = true;
         AZStd::string linkName(link->Name().c_str(), link->Name().size());
-        for (const sdf::Joint* joint : Utils::GetJointsForChildLink(model, linkName, getNestedModelJoints))
+        for (const sdf::Joint* joint : SDFormat::GetJointsForChildLink(model, linkName, getNestedModelJoints))
         {
-            const bool isWheelEntity = Utils::IsWheelHeuristics(model, link);
+            const bool isWheelEntity = SDFormat::IsWheelHeuristics(model, link);
             articulationLinkConfiguration = AddToArticulationConfig(articulationLinkConfiguration, joint, isWheelEntity);
         }
