+++ Gems/ROS2RobotImporter/Code/Source/RobotImporter/URDF/SdfParser.h
@@ -107,4 +107,11 @@ namespace ROS2RobotImporter::SdfParser
     //! @return SDF root object containing parsed <world> or <model> tags
     RootObjectOutcome ParseFromFile(
         AZ::IO::PathView filePath, const sdf::ParserConfig& parserConfig, const SdfAssetBuilderSettings& settings);
+
+    //! Given a set of SdfAssetBuilderSettings, produce an sdf::ParserConfig that can be used by the sdformat library.
+    //! @param settings The input settings to use
+    //! @param baseFilePath The base file getting parsed, which is used to help resolve file paths
+    //! @return The output parser config to use with sdformat.
+    sdf::ParserConfig CreateSdfParserConfigFromSettings(const SdfAssetBuilderSettings& settings, const AZ::IO::PathView& baseFilePath);
+
 } // namespace ROS2RobotImporter::SdfParser
