+++ Gems/ROS2RobotImporter/Code/Source/RobotImporter/Utils/TypeConversions.h
@@ -12,9 +12,12 @@
 #include <AzCore/Math/Quaternion.h>
 #include <AzCore/Math/Transform.h>
 #include <AzCore/Math/Vector3.h>
-#include <RobotImporter/URDF/SdfParser.h>
+#include <gz/math/Color.hh>
+#include <gz/math/Pose3.hh>
+#include <gz/math/Quaternion.hh>
+#include <gz/math/Vector3.hh>
 
-namespace ROS2RobotImporter::Utils
+namespace ROS2RobotImporter::Utils::SDFormat
 {
     //! Common types conversion between gz::math (libsdformat) and AZ formats
     namespace TypeConversions
@@ -24,4 +27,4 @@ namespace ROS2RobotImporter::Utils
         AZ::Color ConvertColor(const gz::math::Color& color);
         AZ::Transform ConvertPose(const gz::math::Pose3d& pose);
     }; // namespace TypeConversions
-} // namespace ROS2RobotImporter::Utils
+} // namespace ROS2RobotImporter::Utils::SDFormat
